backend/requirements.txt
Django==3.2.25
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
django-filter==23.5
django-cors-headers==4.5.0
drf-extra-fields==3.7.0
dotenv==0.9.9
Pillow==10.4.0
PyYAML==6.0
python-dotenv==1.0.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
webcolors==1.11.1